requirements.txt
# ---------- Core scientific stack ----------
numpy>=1.26,<2.0
pandas>=2.1,<3.0
scikit-learn>=1.3,<2.0

# ---------- Deep-learning ----------
torch>=2.1,<3.0

# ---------- Configuration & I/O ----------
pyyaml>=6.0,<7.0

# ---------- Visualisation ----------
matplotlib>=3.8,<4.0

# ---------- Testing ----------
pytest>=7.4,<9.0
